requirements.txt
tensorflow
keras
opencv-python
numpy
pandas
matplotlib
pillow
scikit-learn
imbalanced-learn
streamlit
seaborn